refactor out bimap and add tests
